refactor: drop redundant "lattice" from realtime listener naming
