+++ resources/js/index.ts
@@ -72,7 +72,7 @@ export type {
   UnknownComponent,
   WireNode,
 } from "./core/types";
-export { LatticeListeners } from "./realtime/lattice-listeners";
+export { RealtimeListeners } from "./realtime/listeners";
 export type { ListenerPayload, ListenerVisibility } from "./realtime/types";
 export { columnCell } from "./table/registry";
 export type { ColumnCellArgs, ColumnCellComponent, ColumnRegistry } from "./table/registry";
