+++ resources/js/page.tsx
@@ -2,7 +2,7 @@ import { Head } from "@inertiajs/react";
 import { Renderer } from "@lattice-php/lattice/core/renderer";
 import type { PagePayload } from "@lattice-php/lattice";
 import { cn } from "@lattice-php/lattice/lib/utils";
-import { LatticeListeners } from "@lattice-php/lattice/realtime/lattice-listeners";
+import { RealtimeListeners } from "@lattice-php/lattice/realtime/listeners";
 
 type Props = {
   lattice: PagePayload;
@@ -15,7 +15,7 @@ export default function Page({ lattice }: Props) {
     <>
       <Head title={lattice.title ?? undefined} />
 
-      <LatticeListeners listeners={lattice.listeners} />
+      <RealtimeListeners listeners={lattice.listeners} />
 
       {lattice.container === "centered" ? (
         <main className="min-h-svh bg-lt-bg text-lt-fg">
